requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Cors==4.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
firebase-admin==6.2.0
pyrebase4==4.7.1

# Payment processing
stripe==7.5.0

# AWS
boto3==1.28.64
botocore==1.31.64

# AI
anthropic==0.8.1

# File processing
PyPDF2==3.0.1
python-docx==1.0.1

# Web server
gunicorn==21.2.0
Werkzeug==2.3.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.1.0
Pillow==10.1.0
bleach==6.1.0
markdown==3.5
python-dateutil==2.8.2
pytz==2023.3
